Add Modbus capabilities between main board, motor board and ultrasonic board

+++ code/main_board/src/system/modbus_rtu_crc_table.h
@@ -3,4 +3,4 @@
 #include <stddef.h>
 #include <stdint.h>
 
-uint16_t compute_crc(uint8_t *message, uint8_t len);
+uint16_t compute_crc(uint8_t *message, size_t len);
